# Optimise new submission creation