--- a/app/models/concerns/course/assessment/submission/answers_concern.rb
+++ b/app/models/concerns/course/assessment/submission/answers_concern.rb
@@ -17,15 +17,37 @@ module Course::Assessment::Submission::AnswersConcern
   end
 
   def create_new_answers
-    questions_to_attempt ||= assessment.questions
+    questions_to_attempt ||= assessment.questions.includes(:actable)
     new_answers = questions_to_attempt.not_answered(self).attempt(self)
+    bulk_save_new_answers(new_answers)
+  end
 
-    new_answers.map do |answer|
-      # When there are no existing answers, the first one will be the current_answer.
-      if answer.new_record?
-        answer.current_answer = true
-        answer.save!
+  private
+
+  # Insert new answer records (and its actables) in bulk.
+  #
+  # @param [Array<Course::Assessment::Answer>] new_answers Array of new submission answers
+  # @raise [ActiveRecord::RecordInvalid] If the new answers cannot be saved.
+  # @return[Boolean] If new answers were created.
+  def bulk_save_new_answers(new_answers)
+    # When there are no existing answers, the first one will be the current_answer.
+    # We first filter new_record from the new_answers and assign the current answer flag
+    # below.
+    new_answers_record = new_answers.select(&:new_record?)
+    return false unless new_answers_record.present?
+
+    new_answers_record.each do |new_answer_record|
+      new_answer_record.current_answer = true
+    end
+
+    new_answers_actables = new_answers_record.map(&:actable)
+    new_answers_group_by_actables = new_answers_actables.group_by { |actable| actable.class.to_s }
+
+    ActiveRecord::Base.transaction do
+      new_answers_group_by_actables.each_key do |key|
+        key.constantize.import! new_answers_group_by_actables[key], recursive: true
       end
     end
+    true
   end
 end