--- a/app/services/course/assessment/submission/update_service.rb
+++ b/app/services/course/assessment/submission/update_service.rb
@@ -34,8 +34,7 @@ class Course::Assessment::Submission::UpdateService < SimpleDelegator
   def load_or_create_answers
     return unless @submission.attempting?
 
-    new_answers = questions_to_attempt.includes(:actable).not_answered(@submission).attempt(@submission)
-    new_answers_created = bulk_save_new_answers(new_answers)
+    new_answers_created = @submission.create_new_answers
     @submission.answers.reload if new_answers_created && @submission.answers.loaded?
   end
 
@@ -114,33 +113,6 @@ class Course::Assessment::Submission::UpdateService < SimpleDelegator
 
   def submit_answer_params
     params.require(:answer).permit([:id] + update_answer_type_params)
-  end
-
-  # Insert new answer records (and its actables) in bulk.
-  #
-  # @param [Array<Course::Assessment::Answer>] new_answers Array of new submission answers
-  # @raise [ActiveRecord::RecordInvalid] If the new answers cannot be saved.
-  # @return[Boolean] If new answers were created.
-  def bulk_save_new_answers(new_answers)
-    # When there are no existing answers, the first one will be the current_answer.
-    # We first filter new_record from the new_answers and assign the current answer flag
-    # below.
-    new_answers_record = new_answers.select(&:new_record?)
-    return false unless new_answers_record.present?
-
-    new_answers_record.each do |new_answer_record|
-      new_answer_record.current_answer = true
-    end
-
-    new_answers_actables = new_answers_record.map(&:actable)
-    new_answers_group_by_actables = new_answers_actables.group_by { |actable| actable.class.to_s }
-
-    ActiveRecord::Base.transaction do
-      new_answers_group_by_actables.each_key do |key|
-        key.constantize.import! new_answers_group_by_actables[key], recursive: true
-      end
-    end
-    true
   end
 
   def questions_to_attempt