When a user attempts a submission for the first time, submission_questions and answers are created and inserted 1 by 1 to the DB resulting in N+1 issue. To solve this, we now use bulk insert instead.

In addition, a new submission of a randomized assessment would create answers and submission_questions for all questions from the assessment. Now, only assigned questions based on the question bundle assignmenets are used.

Improvement comparisons (create new attempt on courses/1745/assessments/30193):
- Local (before: ~2600ms, after: ~1300ms)
- Staging (before: ~700ms, after: ~ms)
- Production (before: ~1000ms, after: ~ms)

In detail (local):
Before
<img width="559" alt="image" src="https://user-images.githubusercontent.com/42570513/149257705-52f2c211-3a97-40f8-9c67-e511ce8f9cb7.png">

After
![image](https://user-images.githubusercontent.com/42570513/149258425-2d7dff61-7875-4367-9f70-b2745fb86754.png)