@ekowidianto 2022-01-05
@zhuhanming Asking your opinion for this issue if we could do it this way or any other alternatives. With the bulk insert for the submission_question, we could cut another ~20% of the time to execute the whole thing. Minor downside is that since submission_question is an actable of discussion post, we need to explicitly do another bulk insert for the post. CMIIW but currently there is no way to do bulk insert with association.

It is also possible to do the same thing for the answers creation ([def load_or_create_answers](https://github.com/Coursemology/coursemology2/blob/3c84ce4f9250511835fbe1f7a74909e43b1896d4/app/services/course/assessment/submission/update_service.rb#L34)). I think it may be similar to the submission_question above, where we need to bulk insert the answer actables > then bulk insert answers, but it gets a little bit trickier here since some answer actables have another association to save (eg for programming answer actables > programming files). So it gets a little bit more manual here to figure out all the associations that are saved and we need to write the bulk insert separately for each table. Also, since we have 6 different types of the answer actables, there will be different types of question in an assessment and the bulk insert will need to be done separately (best case all same then 1 bulk insert or worst case all different then nothing happens). What do you think about this? Or we could just stick to the current way of saving (i.e. individual save)

TLDR; working with bulk insert for MTI is tricky.