@ekowidianto 2022-01-13
Comparisons for randomized assessment (2040S Final)

**Before**
<img width="559" alt="image" src="https://user-images.githubusercontent.com/42570513/149257360-dd2ae1dc-52f3-47f4-8742-94b75b4bdab6.png">

**After**
<img width="567" alt="image" src="https://user-images.githubusercontent.com/42570513/149256926-3bff896f-d69a-4bc3-a260-21e63431631e.png">